apps/api/pyproject.toml
[project]
name = "portfolioai-api"
version = "0.1.0"
description = "PortfolioAI Backend API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "supabase>=2.0.0",
    "pydantic[email]>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "yfinance>=0.2.36",
    "anthropic>=0.39.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]
